implemented get and put
we still need expression to get digits
for some reason with the test code from proj 3 sheet doesnt work in our syntax analyzer? i think its because the inputQ doesnt get popped off in the correct way.

but if you run code with out running the syntax analyzer it performs the correct syntax we just need to add the digits to expression.